Refactor: move integration tests (http requests to the app and db usage) into the tests folder. Split src into a library and a binary.
