+++ src/conduit/articles.rs
@@ -10,8 +10,6 @@ use crate::schema::articles;
 
 type Repo = db::Repo<PgConnection>;
 
-// joinable!(articles -> users (user_id));
-
 #[derive(Default, Deserialize, Debug)]
 pub struct ArticleQuery {
     pub author: Option<String>,
@@ -67,61 +65,4 @@ pub fn find_one(repo: &Repo, slug_value: &str) -> Result<Article, Error> {
             .select(articles::all_columns())
             .first(&conn)
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::conduit::users;
-    use crate::models::{NewArticle, NewUser};
-    use crate::test_helpers::test_server::get_repo;
-
-    use crate::auth::encode_token;
-    use futures_executor::ThreadPool;
-    use uuid::Uuid;
-    // use tokio_async_await_test::async_test;
-
-    // #[async_test]
-    // async fn test_list_articles() {
-    //     let repo = Repo::new();
-
-    //     let users =  create_users(&repo, 5).await ;
-    //     let _articles =  create_articles(&repo, users);
-    //     let results =
-    //         find(repo.clone(), Default::default()).await.expect("Failed to get articles");
-
-    //     assert_eq!(results.len(), 5);
-    // }
-
-    #[test]
-    fn insert_and_retrieve_article() {
-        let runtime = ThreadPool::new().unwrap();
-        runtime.spawn_ok(async move {
-            let repo = get_repo();
-            let slug = "my_slug".to_string();
-
-            let user_id = Uuid::new_v4();
-            let token = encode_token(user_id);
-            let user = NewUser {
-                username: "my_user".into(),
-                email: "my_email@hotmail.com".into(),
-                password: "somepass".into(),
-                id: user_id,
-                token,
-            };
-            let user = users::insert(&repo, user).unwrap();
-
-            let article = NewArticle {
-                title: "My article".into(),
-                slug: slug.clone(),
-                description: "My article description".into(),
-                body: "ohoh".into(),
-                user_id: user.id,
-            };
-            let expected_article = insert(&repo, article).unwrap();
-
-            let retrieved_article = find_one(&repo, &slug).unwrap();
-            assert_eq!(expected_article, retrieved_article);
-        })
-    }
 }
