+++ src/main.rs
@@ -1,67 +1,14 @@
-#[macro_use]
-extern crate diesel;
-
-mod auth;
-mod conduit;
-mod configuration;
-mod db;
-mod middleware;
-mod models;
-mod schema;
-mod web;
-
-#[cfg(test)]
-mod test_helpers;
-
-use crate::configuration::Settings;
 use async_std::task::block_on;
-use diesel::PgConnection;
-use tide::{IntoResponse, Response, Server};
-
-type Repo = db::Repo<PgConnection>;
-
-pub fn result_to_response<T: IntoResponse, E: IntoResponse>(r: Result<T, E>) -> Response {
-    match r {
-        Ok(r) => r.into_response(),
-        Err(r) => {
-            let res = r.into_response();
-            if res.status().is_success() {
-                panic!(
-                    "Attempted to yield error response with success code {:?}",
-                    res.status()
-                )
-            }
-            res
-        }
-    }
-}
-
-pub fn set_routes(mut app: Server<Repo>) -> Server<Repo> {
-    app.at("/api").nest(|api| {
-        api.at("/user")
-            .get(|req| async move { result_to_response(web::users::get_current_user(req).await) })
-            .put(|req| async move { result_to_response(web::users::update_user(req).await) });
-        api.at("/users")
-            .post(|req| async move { result_to_response(web::users::register(req).await) });
-        api.at("/users/login")
-            .post(|req| async move { result_to_response(web::users::login(req).await) });
-        api.at("/articles")
-            .get(|req| async move { result_to_response(web::articles::list_articles(req).await) });
-        api.at("/articles/:slug")
-            .get(|req| async move { result_to_response(web::articles::get_article(req).await) });
-    });
-    app
-}
+use realworld_tide::configuration::Settings;
+use realworld_tide::db::Repo;
+use realworld_tide::web::get_app;
 
 fn main() -> Result<(), std::io::Error> {
     let settings = Settings::new().expect("Failed to load configuration");
     env_logger::init();
 
     let state = Repo::new(&settings.database.connection_string());
-    let mut app = Server::with_state(state);
-    app.middleware(tide::middleware::RequestLogger::new());
-    app.middleware(crate::middleware::JwtMiddleware::new());
-    app = set_routes(app);
+    let app = get_app(state);
     let address = format!(
         "{}:{}",
         settings.application.host, settings.application.port
