+++ src/web/mod.rs
@@ -1,8 +1,12 @@
+pub mod app;
 pub mod articles;
 pub mod users;
+
 use http::status::StatusCode;
 use log::error;
 use tide::Error;
+
+pub use app::get_app;
 
 pub fn diesel_error(e: &diesel::result::Error) -> Error {
     error!("{}", e);
